[FEATURE] Fixes #94: Significant refactoring of the attachment procesing subsystem.
This lays the groundwork for future multiple document type functionality by allowing
multiple processors to act on one attachment and to generate attachment variants. This
is a breaking change that will require not only migrations to be run, but re-transcoding
of all attachments. The re-transcoding mechanism will come in a subsequent revision.